refactor: 지적사항 업데이트
